codes/chapter5/requirements.txt
attrs==21.4.0
base58==2.1.1
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
ecdsa==0.17.0
Flask==2.0.3
Flask-SocketIO==5.1.1
fonttools==4.31.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.4.0
MarkupSafe==2.1.0
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
six==1.16.0
socketIO-client==0.7.2
tomli==2.0.1
urllib3==1.26.9
websocket-client==1.3.1
Werkzeug==2.0.3
